# public/swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "UserCreateAccessToken": {
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": ["accessToken"],
        "type": "object"
      },
      "Pick_UserCreateResponse.Exclude_keyofUserCreateResponse.id-or-isAdmin__": {
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "email", "password"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_UserCreateResponse.id-or-isAdmin_": {
        "$ref": "#/components/schemas/Pick_UserCreateResponse.Exclude_keyofUserCreateResponse.id-or-isAdmin__",
        "description": "Construct a type with the properties of T except for those in type K."
      },
      "UserCreateRequest": {
        "$ref": "#/components/schemas/Omit_UserCreateResponse.id-or-isAdmin_"
      }
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "collections_server",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "fayzikuloff.b@gmail.com"
    }
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "post": {
        "operationId": "Create",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateAccessToken"
                }
              }
            }
          }
        },
        "tags": ["users"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "GetAll",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "userId": {
                        "type": "number",
                        "format": "double"
                      }
                    },
                    "required": ["userId"],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["users"],
        "security": [],
        "parameters": []
      }
    },
    "/api/users/{userId}": {
      "get": {
        "operationId": "GetOne",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userId": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["userId"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["users"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userId": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["userId"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["users"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "delete": {
        "operationId": "Delete",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "userId": {
                      "type": "number",
                      "format": "double"
                    }
                  },
                  "required": ["userId"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["users"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}